unit test and refactoring
