+++ package.json
@@ -2,12 +2,16 @@
   "name": "node-starter-v2",
   "version": "1.0.0",
   "scripts": {
+    "local": "yarn start:local",
+    "dev": "yarn start:dev",
+    "test": "yarn start:test",
     "build": "tsoa spec-and-routes && tsc",
-    "start": "yarn start:dev",
-    "test:integration": "yarn build && yarn env:test mocha --timeout 35000 -r ts-node/register ./src/**/*.spec.ts --exit",
-    "start:local": "yarn env:local concurrently \"nodemon\" \"nodemon -x tsoa spec-and-routes && tsc \"",
-    "start:dev": "yarn env:dev ts-node-dev ./src/server.ts",
-    "start:test": "test:integration",
+    "lint": "yarn build && tslint -c \"./tslint.json\" -p \"./tsconfig.json\"",
+    "test:unit": "yarn build && yarn env:test nyc mocha --timeout 10000 -r ts-node/register ./src/**/test/unit/*.spec.ts",
+    "test:integration": "yarn build && yarn env:test mocha --timeout 35000 -r ts-node/register ./src/**/test/integration/*.spec.ts --exit",
+    "start:local": "yarn build && yarn env:local concurrently \"nodemon\" \"nodemon -x tsoa spec-and-routes && tsc \"",
+    "start:dev": "yarn build && yarn env:dev ts-node-dev ./src/server.ts",
+    "start:test": "yarn lint && yarn test:unit && yarn test:integration",
     "env:local": "cross-env NODE_ENV=local",
     "env:dev": "cross-env NODE_ENV=development",
     "env:test": "cross-env NODE_ENV=test"
@@ -48,6 +52,7 @@
     "supertest": "^6.2.4",
     "ts-node": "^10.8.2",
     "ts-node-dev": "^2.0.0",
+    "tslint": "^6.1.3",
     "typescript": "^4.7.4"
   }
 }
