+++ src/domain/users/test/integration/users.spec.ts
@@ -1,7 +1,7 @@
 import { expect } from 'chai';
 import supertest from 'supertest';
 
-import { app } from "../app";
+import { app } from "../../../../app";
 const entityName: string = 'user';
 
 describe(`users`, () => {
