+++ src/utils/Logger.ts
@@ -1,4 +1,4 @@
-import constants from './constants';
+import constants from './../config/constants';
 import * as logger from 'winston';
 
 const date = new Date();
@@ -17,11 +17,6 @@ logger.configure({
 export class Logger {
   public static readonly shouldLog: boolean = constants.environment !== 'test';
   public static readonly console = logger;
-
-  public constructor() {
-    console.log(constants.environment);
-  }
-
 
   public static log(...args: any[]): void {
     if (Logger.shouldLog) Logger.console.debug(Logger.formatArgs(args));
@@ -43,9 +38,12 @@ export class Logger {
     if (Logger.shouldLog) Logger.console.verbose(Logger.formatArgs(args));
   }
 
+  public constructor() {
+    console.log(constants.environment);
+  }
+
   private static formatArgs(args: any[]): string {
     if (args.length <= 1) args = args[0];
     return JSON.stringify(args, null, 4);
   }
-
 }
