+++ tsoa.json
@@ -1,13 +1,13 @@
 {
   "entryFile": "src/app.ts",
   "noImplicitAdditionalProperties": "throw-on-extras",
-  "controllerPathGlobs": ["src/**/*Controller.ts"],
+  "controllerPathGlobs": ["src/**/**/*Controller.ts"],
   "spec": {
     "outputDirectory": "build",
     "specVersion": 3
   },
   "routes": {
     "routesDir": "build",
-    "iocModule": "src/ioc"
+    "iocModule": "src/config/ioc"
   }
 }
